+++ AppCenterReactNativeShared/ios/AppCenterReactNativeShared/AppCenterReactNativeShared.m
@@ -5,14 +5,6 @@
 #import <AppCenter/MSAppCenter.h>
 #import <AppCenter/MSWrapperSdk.h>
 
-// Сonvince the compiler that this private class exists and implemented.
-@interface MSAuthTokenContext
-
-+ (instancetype)sharedInstance;
-- (void)preventResetAuthTokenAfterStart;
-
-@end
-
 @implementation AppCenterReactNativeShared
 
 static NSString *const kAppCenterSecretKey = @"AppSecret";
@@ -60,15 +52,6 @@ static NSDictionary *configuration;
         [MSAppCenter configure];
       } else {
         [MSAppCenter configureWithAppSecret:appSecret];
-      }
-
-      /*
-       * When startAutomatically flag is set to true, every service (analytics/auth/crashes/etc.)
-       * will be started by separate AppCenter.start call. If Auth module is used,
-       * call preventResetAuthTokenAfterStart to avoid resetting the auth token.
-       */
-      if (NSClassFromString(@"MSAuth")) {
-        [[MSAuthTokenContext sharedInstance] preventResetAuthTokenAfterStart];
       }
     }
   }