+++ DemoApp/android/app/src/main/java/com/demoapp/MainApplication.java
@@ -15,9 +15,7 @@ import com.imagepicker.ImagePickerPackage;
 import com.microsoft.appcenter.AppCenter;
 import com.microsoft.appcenter.reactnative.analytics.AppCenterReactNativeAnalyticsPackage;
 import com.microsoft.appcenter.reactnative.appcenter.AppCenterReactNativePackage;
-import com.microsoft.appcenter.reactnative.auth.AppCenterReactNativeAuthPackage;
 import com.microsoft.appcenter.reactnative.crashes.AppCenterReactNativeCrashesPackage;
-import com.microsoft.appcenter.reactnative.data.AppCenterReactNativeDataPackage;
 import com.microsoft.appcenter.reactnative.push.AppCenterReactNativePushPackage;
 import com.reactnativecommunity.asyncstorage.AsyncStoragePackage;
 import com.rnfs.RNFSPackage;
@@ -39,12 +37,10 @@ public class MainApplication extends Application implements ReactApplication {
         protected List<ReactPackage> getPackages() {
             return Arrays.asList(
                     new MainReactPackage(),
-                    new AppCenterReactNativeDataPackage(MainApplication.this),
                     new RNGestureHandlerPackage(),
                     new ImagePickerPackage(),
                     new RNFSPackage(),
                     new DemoAppNativePackage(MainApplication.this),
-                    new AppCenterReactNativeAuthPackage(MainApplication.this),
                     new AppCenterReactNativePushPackage(MainApplication.this),
                     new AppCenterReactNativeCrashesPackage(MainApplication.this, getResources().getString(R.string.appCenterCrashes_whenToSendCrashes)),
                     new AppCenterReactNativeAnalyticsPackage(MainApplication.this, getResources().getString(R.string.appCenterAnalytics_whenToEnableAnalytics)),