+++ DemoApp/ios/DemoApp/AppDelegate.m
@@ -11,9 +11,7 @@
  */
 
 #import "AppDelegate.h"
-#import <AppCenterReactNativeData/AppCenterReactNativeData.h>
 #import <AppCenterReactNative/AppCenterReactNative.h>
-#import <AppCenterReactNativeAuth/AppCenterReactNativeAuth.h>
 #import <AppCenterReactNativePush/AppCenterReactNativePush.h>
 #import <AppCenterReactNativeCrashes/AppCenterReactNativeCrashes.h>
 #import <AppCenterReactNativeAnalytics/AppCenterReactNativeAnalytics.h>
@@ -41,10 +39,8 @@
 
   [AppCenterReactNative register];                                   // Initialize AppCenter
   [AppCenterReactNativeAnalytics registerWithInitiallyEnabled:true]; // Initialize AppCenter Analytics
-  [AppCenterReactNativeAuth register];                               // Initialize AppCenter Auth
   [AppCenterReactNativeCrashes register];                            // Initialize AppCenter Crashes
   [AppCenterReactNativePush register];                               // Initialize AppCenter Push
-  [AppCenterReactNativeData register];                               // Initialize AppCenter Data
 
   RCTBridge *bridge = [[RCTBridge alloc] initWithDelegate:self launchOptions:launchOptions];
   RCTRootView *rootView = [[RCTRootView alloc] initWithBridge:bridge moduleName:@"DemoApp" initialProperties:nil];