+++ DemoApp/upgrade-appcenter-npm-packages.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 set -e
 
-npm uninstall --save appcenter appcenter-analytics appcenter-auth appcenter-crashes appcenter-push appcenter-data
+npm uninstall --save appcenter appcenter-analytics appcenter-crashes appcenter-push
 
-npm install --save appcenter appcenter-analytics appcenter-auth appcenter-crashes appcenter-push appcenter-data
+npm install --save appcenter appcenter-analytics appcenter-crashes appcenter-push
 
 pod repo update