+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # App Center SDK for React Native Change Log
 
+## Version 3.0.0 (Under development)
+
+### App Center Auth
+
+App Center Auth is [retired](https://aka.ms/MBaaS-retirement-blog-post) and has been removed from the SDK.
+
+### App Center Data
+
+App Center Data is [retired](https://aka.ms/MBaaS-retirement-blog-post) and has been removed from the SDK.
+
+___
+
 ## Version 2.6.2 (Under development)
 
 ### App Center